refactor(expand): WordPolicy names the word rule, classify becomes a package
The word rule is now an explicit WordPolicy enum (route/types) derived
from the consumer: SHELL words are always shell-resolved and spec hints
ignored; MOUNT words classify by spec with patterns kept for pushdown.
spec_for_command owns the cwd-mount/shared-SPECS lookup as the single
seam for the default-mount split, asserting that mount_for is total
instead of defensively catching. classify splits into a package
(heuristic/path/parts) so the shape heuristics have their own module.
TS mirrors, and posixNormpath/shlexSplit move from classify to utils.
