+++ python/mirage/workspace/expand/argv.py
@@ -18,16 +18,16 @@ from dataclasses import dataclass
 
 import tree_sitter
 
-from mirage.commands.spec import SPECS
 from mirage.commands.spec.types import OperandKind
 from mirage.shell.call_stack import CallStack
 from mirage.types import PathSpec
 from mirage.workspace.expand.classify import classify_parts
 from mirage.workspace.expand.globs import resolve_globs
 from mirage.workspace.expand.parts import expand_parts
-from mirage.workspace.expand.spec_hints import spec_word_kinds
+from mirage.workspace.expand.spec_hints import (spec_for_command,
+                                                spec_word_kinds)
 from mirage.workspace.mount import MountRegistry
-from mirage.workspace.route import SHELL_CONSUMERS, route
+from mirage.workspace.route import WordPolicy, route, word_policy
 from mirage.workspace.session import Session
 
 
@@ -99,23 +99,10 @@ async def expand_argv(
         return Argv(name="", args=(), operands=())
     name = expanded[0]
 
-    # Word policy follows the consumer (route table): shell consumers get
-    # bash semantics, where globs expand no matter what the command does
-    # with its arguments, so spec TEXT kinds must not suppress them. Only
-    # mount commands classify by spec.
-    is_shell = route(name, session, registry) in SHELL_CONSUMERS
+    policy = word_policy(route(name, session, registry))
     word_kinds: list[OperandKind | None] | None = None
-    if not is_shell:
-        try:
-            cwd_mount = registry.mount_for(session.cwd)
-        except ValueError:
-            cwd_mount = None
-        spec = cwd_mount.spec_for(name) if cwd_mount else None
-        if spec is None:
-            # The cwd mount may not know the command, or cwd is unmounted;
-            # fall back to the shared SPECS so classification still gets
-            # per-position kinds instead of shape heuristics.
-            spec = SPECS.get(name)
+    if policy is WordPolicy.MOUNT:
+        spec = spec_for_command(name, registry, session.cwd)
         if spec:
             word_kinds = spec_word_kinds(spec, expanded[1:])
 
@@ -124,9 +111,10 @@ async def expand_argv(
                                 session.cwd,
                                 word_kinds=word_kinds)
     # A glob word is resolved by whoever consumes it, exactly once:
-    # shell consumers get matches here; mount commands keep patterns for
-    # backend pushdown; unknown names fail without touching backends.
-    if is_shell:
+    # WordPolicy.SHELL words get matches here; mount commands keep
+    # patterns for backend pushdown; unknown names fail without
+    # touching backends.
+    if policy is WordPolicy.SHELL:
         words = await resolve_globs(classified, registry)
     else:
         words = classified
