+++ python/tests/workspace/expand/test_spec_hints.py
@@ -17,10 +17,30 @@ import pytest
 from mirage import MountMode, RAMResource, Workspace
 from mirage.commands.spec import SPECS
 from mirage.commands.spec.types import OperandKind
-from mirage.workspace.expand.spec_hints import spec_word_kinds
+from mirage.workspace.expand.spec_hints import (spec_for_command,
+                                                spec_word_kinds)
 
 PATH = OperandKind.PATH
 TEXT = OperandKind.TEXT
+
+
+def test_spec_for_command_prefers_cwd_mount():
+    ws = Workspace({"/": RAMResource()}, mode=MountMode.WRITE)
+    mount = ws._registry.mount_for("/")
+    spec = spec_for_command("grep", ws._registry, "/")
+    assert spec is mount.spec_for("grep")
+
+
+def test_spec_for_command_falls_back_to_shared_specs():
+    ws = Workspace({"/": RAMResource()}, mode=MountMode.WRITE)
+    mount = ws._registry.mount_for("/")
+    name = next(n for n in SPECS if mount.spec_for(n) is None)
+    assert spec_for_command(name, ws._registry, "/") is SPECS[name]
+
+
+def test_spec_for_command_unknown_name_is_none():
+    ws = Workspace({"/": RAMResource()}, mode=MountMode.WRITE)
+    assert spec_for_command("no-such-command", ws._registry, "/") is None
 
 
 def test_basic_grep_pattern_and_path():
