+++ typescript/packages/core/src/index.ts
@@ -245,7 +245,8 @@ export { fmtGeneric } from './commands/builtin/generic/fmt.ts'
 export { headStream } from './commands/builtin/generic/head.ts'
 export { basenameFn } from './commands/builtin/generic/basename.ts'
 export { dirnameFn } from './commands/builtin/generic/dirname.ts'
-export { gnuBasename, gnuDirname, norm, parent } from './utils/path.ts'
+export { gnuBasename, gnuDirname, norm, parent, posixNormpath } from './utils/path.ts'
+export { shlexSplit } from './utils/shlex.ts'
 export { detectFileType, FILE_MIME_MAP, formatFileResult } from './commands/builtin/file_helper.ts'
 export {
   type AggregateResult,
@@ -512,9 +513,7 @@ export {
   expandParts,
   type ExecuteFn,
   lookupVar,
-  posixNormpath,
   safeEval,
-  shlexSplit,
   type TSNodeLike,
   unescapePath,
 } from './workspace/expand/index.ts'
