+++ typescript/packages/core/src/workspace/executor/builtins/dirs.ts
@@ -17,7 +17,7 @@ import { IOResult } from '../../../io/types.ts'
 import type { PathSpec } from '../../../types.ts'
 import { FileType } from '../../../types.ts'
 import { CycleError, MAX_SYMLINK_HOPS, resolveSymlinks } from '../../../utils/path.ts'
-import { posixNormpath } from '../../expand/classify.ts'
+import { posixNormpath } from '../../../utils/path.ts'
 import type { Session } from '../../session/session.ts'
 import { changeDir } from '../../session/shell_dirs.ts'
 import { ExecutionNode } from '../../types.ts'
