+++ typescript/packages/core/src/workspace/expand/classify/heuristic.test.ts
@@ -13,18 +13,10 @@
 // ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
 import { describe, expect, it } from 'vitest'
-import { OperandKind } from '../../commands/spec/types.ts'
-import type { Resource } from '../../resource/base.ts'
-import { MountMode, PathSpec } from '../../types.ts'
-import { MountRegistry } from '../mount/registry.ts'
-import {
-  classifyBarePath,
-  classifyParts,
-  classifyWord,
-  posixNormpath,
-  shlexSplit,
-  unescapePath,
-} from './classify.ts'
+import type { Resource } from '../../../resource/base.ts'
+import { MountMode, PathSpec } from '../../../types.ts'
+import { MountRegistry } from '../../mount/registry.ts'
+import { classifyWord, unescapePath } from './heuristic.ts'
 
 class StubResource implements Resource {
   readonly kind = 'stub'
@@ -39,46 +31,6 @@ class StubResource implements Resource {
 function setup(): MountRegistry {
   return new MountRegistry({ '/ram': new StubResource() }, MountMode.WRITE)
 }
-
-describe('posixNormpath', () => {
-  it('normalizes ..', () => {
-    expect(posixNormpath('/a/b/../c')).toBe('/a/c')
-  })
-
-  it('collapses redundant slashes', () => {
-    expect(posixNormpath('/a//b')).toBe('/a/b')
-  })
-
-  it('drops trailing slash', () => {
-    expect(posixNormpath('/a/b/')).toBe('/a/b')
-  })
-
-  it('handles relative paths', () => {
-    expect(posixNormpath('a/./b/../c')).toBe('a/c')
-  })
-
-  it('empty string becomes .', () => {
-    expect(posixNormpath('')).toBe('.')
-  })
-})
-
-describe('shlexSplit', () => {
-  it('splits whitespace-separated words', () => {
-    expect(shlexSplit('a b c')).toEqual(['a', 'b', 'c'])
-  })
-
-  it('backslash escapes the next char', () => {
-    expect(shlexSplit('hello\\ world')).toEqual(['hello world'])
-  })
-
-  it('single quotes preserve everything', () => {
-    expect(shlexSplit("'a b'")).toEqual(['a b'])
-  })
-
-  it('double quotes support escape sequences', () => {
-    expect(shlexSplit('"a\\"b"')).toEqual(['a"b'])
-  })
-})
 
 describe('unescapePath', () => {
   it('strips backslash escapes from paths', () => {
@@ -139,45 +91,5 @@ describe('classifyWord — relative paths', () => {
   it('leaves bare glob (like *) as text — could be a command arg', () => {
     const reg = setup()
     expect(classifyWord('*', reg, '/ram')).toBe('*')
-  })
-})
-
-describe('classifyBarePath', () => {
-  it('forces classification of a bare filename under cwd', () => {
-    const reg = setup()
-    const r = classifyBarePath('file.txt', reg, '/ram')
-    if (!(r instanceof PathSpec)) throw new Error('expected PathSpec')
-    expect(r.virtual).toBe('/ram/file.txt')
-  })
-})
-
-describe('classifyParts', () => {
-  it('first arg (command name) is never classified', () => {
-    const reg = setup()
-    const out = classifyParts(['cat', '/ram/x'], reg, '/')
-    expect(out[0]).toBe('cat')
-    expect(out[1]).toBeInstanceOf(PathSpec)
-  })
-
-  it('TEXT kind forces plain-text classification', () => {
-    const reg = setup()
-    const out = classifyParts(['cat', '/ram/x'], reg, '/', [OperandKind.TEXT])
-    expect(out[1]).toBe('/ram/x')
-  })
-
-  it('PATH kind forces bare-path classification', () => {
-    const reg = setup()
-    const out = classifyParts(['cat', 'file.txt'], reg, '/ram', [OperandKind.PATH])
-    expect(out[1]).toBeInstanceOf(PathSpec)
-  })
-
-  it('duplicate word classifies per slot', () => {
-    const reg = setup()
-    const out = classifyParts(['grep', '*.txt', '*.txt'], reg, '/ram', [
-      OperandKind.TEXT,
-      OperandKind.PATH,
-    ])
-    expect(out[1]).toBe('*.txt')
-    expect(out[2]).toBeInstanceOf(PathSpec)
   })
 })
