+++ typescript/packages/core/src/workspace/expand/redirects.ts
@@ -16,7 +16,7 @@ import type { CallStack } from '../../shell/call_stack.ts'
 import { Redirect, RedirectKind } from '../../shell/types.ts'
 import type { MountRegistry } from '../mount/registry.ts'
 import type { Session } from '../session/session.ts'
-import { classifyBarePath } from './classify.ts'
+import { classifyBarePath } from './classify/index.ts'
 import { expandNode } from './node.ts'
 import type { ExecuteFn } from './node.ts'
 import type { TSNodeLike } from './variable.ts'
