+++ typescript/packages/core/src/workspace/node/provision_node.ts
@@ -36,7 +36,7 @@ import { Precision, ProvisionResult } from '../../provision/types.ts'
 import { rollupList, rollupPipe } from '../../provision/rollup.ts'
 import { PathSpec } from '../../types.ts'
 import type { TSNodeLike } from '../expand/variable.ts'
-import { classifyParts } from '../expand/classify.ts'
+import { classifyParts } from '../expand/classify/index.ts'
 import type { ExecuteFn } from '../expand/node.ts'
 import { expandAndClassify, expandParts } from '../expand/parts.ts'
 import { expandRedirects } from '../expand/redirects.ts'
